chore: add ts linting
