+++ .github/workflows/ci.yml
@@ -25,7 +25,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: ./.github/actions/setup
-      - run: pnpm compile
+      - run: pnpm lint:ts
 
   eslint:
     if: github.head_ref != 'changeset-release/main'
