+++ package.json
@@ -21,7 +21,7 @@
     "storybook:build:prod": "pnpm storybook:build --docs",
     "test:storybook": "test-storybook --config-dir storybook",
     "jest:debug": "node --inspect-brk node_modules/.bin/jest --runInBand --no-cache",
-    "compile": "tsc && pnpm pkg:aio compile",
+    "lint:ts": "tsc --noEmit && pnpm turbo lint:ts",
     "eslint-config": "pnpm eslint -c .eslintrc.js --max-warnings=0",
     "lint": "pnpm lint:js && pnpm lint:md && pnpm lint:format && pnpm lint:styles",
     "lint:js": "pnpm eslint-config '**/*.{ts,tsx,mjs,js}'",
