+++ packages/components/package.json
@@ -29,6 +29,7 @@
     "build": "pnpm package-bundler build-shared-ui && pnpm build:global-styles",
     "build:global-styles": "postcss ./styles/global.css --output dist/styles.css",
     "clean": "rm -rf dist",
+    "lint:ts": "tsc --noEmit",
     "test": "FORCE_COLOR=1 jest",
     "test:ci": "pnpm test -- --ci",
     "update-icons": "./src/Icon/bin/update-icons.sh",
