+++ packages/components/types.d.ts
@@ -7,11 +7,3 @@ declare module "*.scss" {
   const classes: Record<string, string>
   export default classes
 }
-
-declare module "*.icon.svg" {
-  const content: {
-    id: string
-    viewBox: string
-  }
-  export default content
-}
