+++ packages/design-tokens/docs/DocsComponents.tsx
@@ -4,8 +4,7 @@ import React from "react"
 import { Unstyled } from "@storybook/blocks"
 import { toCustomMediaQueriesArray } from "object-to-css-variables"
 import Highlight from "react-highlight"
-import { Card } from "~components/Card"
-import { Tab, TabList, TabPanel, TabPanels, Tabs } from "~components/Tabs"
+import { Card, Tab, TabList, TabPanel, TabPanels, Tabs } from "../../components"
 import { tokens } from "../src/js"
 import { makeCssVariableDefinitionsMap } from "../src/lib/makeCssVariableDefinitionsMap"
 import animationSass from "!!raw-loader!../sass/animation.scss"
