+++ packages/design-tokens/package.json
@@ -22,6 +22,7 @@
   "license": "MIT",
   "sideEffects": false,
   "scripts": {
+    "lint:ts": "tsc --noEmit",
     "test": "FORCE_COLOR=1 jest",
     "test:ci": "pnpm test -- --ci",
     "build": "pnpm clean && pnpm build:json && pnpm build:less && pnpm build:sass && pnpm build:ts",
