+++ packages/package-bundler/package.json
@@ -9,7 +9,8 @@
   },
   "scripts": {
     "build": "rm -rf dist && tsc --build",
-    "prepare": "ts-patch install -s"
+    "prepare": "ts-patch install -s",
+    "lint:ts": "tsc --noEmit"
   },
   "files": [
     "dist",
