+++ packages/package-bundler/tsconfig.json
@@ -11,7 +11,7 @@
     "strict": true,
     "allowJs": false,
     "outDir": "dist",
-    "target": "ESNext"
+    "target": "ESNext",
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules"]
