+++ packages/tailwind/package.json
@@ -4,7 +4,8 @@
   "description": "Kaizen Tailwind presets",
   "scripts": {
     "build": "pnpm clean && tsc --project tsconfig.dist.json",
-    "clean": "rimraf -g '**/*.d.ts' '**/*.js' '**/*.map'"
+    "clean": "rimraf -g '**/*.d.ts' '**/*.js' '**/*.map'",
+    "lint:ts": "tsc --noEmit"
   },
   "repository": {
     "type": "git",
